Templates/Completed routes
Added remaining jinja2 templating needed display required query/result information for this assignment.
Added 3 variables into the render_template()s on the server.py to display an error of duplication of a email already in the database, another to show success of the newly added email, and a for loop showing all of the emails currently in the database.
Added 3 queries to server.py allow all of the above to work with no issues.