Date/stying/delete
The datetime listed when a specific email was created was added to the server.py to be displayed on success.html next to each email.
Added styling for each event: Entering in a valid email or not. Styling was added to style.css and classes placed on to the html files through the variables listed on server.py:  resultString, fail.